+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # BetterDisplay Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [Change resolutions] - 2025-03-01
+
+* Added the ability to change resolutions.
+
+## [Initial Version] - 2024-02-28